Get SQL Server points when we pan / zoom the map
Whenever we zoom or pan on the map we fire an AJAX call to get points
for the current map bounds from SQL Server.

There is a setTimeout in here so we only call the back end if there has
been no map movement for 1 second.

Added in new Loading and "message" popups.