Added a LoadFromCircle method
Added a LoadFromCircle method to see if it would solve the projection
problems where the returned shape was a so curved most of the points
were off the visible screen (even though we sent the correct rectangle).
It has kind of worked but seems to be a lot slower.

Added some comments.