Using 8 point polygon for accurate boundary
Changed code to generate and use an 8 point polygon to offset the
curvature of the earth when the 4 point polygon was used in SQL Server.

Added .NET style commenting to the JS file

Refactored the Model classes

In order to return the ID of each point (so we can get popup info for
each) I had to return each individual POINT from the database instead of
1 large MULTIPOINT WKT string. This means we now need to loop through
all points in the JS to bind a popup and add them to the cluster layer.

Because of the above I have cut the number of points being returned to
2000 so we don't block the UI for too long when looping